backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0

# RAG & Embeddings
langchain==0.1.4
langchain-core==0.1.20
langchain-openai==0.0.5
langchain-community==0.0.16
langchain-groq==0.1.3
chromadb==0.4.22
pypdf==3.17.4
python-docx==1.1.0

# FREE LLM APIs
groq==0.9.0
httpx==0.27.0
httpcore==1.0.5

# FREE Embeddings
sentence-transformers==2.3.1

# Utilities
aiofiles==23.2.1
numpy==1.26.3